.fixed some warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@6330 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Sql/SqlCase.cpp
@@ -321,7 +321,7 @@ String SqlFormat(const char *s, int l)
 
 String SqlFormat(const char *s)
 {
-	return SqlFormat(s, strlen(s));
+	return SqlFormat(s, (int)strlen(s));
 }
 
 String SqlFormat(const String& x)
